Extract config push into pushLocal/pushRemoteAdmin + localServer const
Follow-up cleanup to the in-process local push. updateServer no longer
embeds the admin-HTTP request building or the "is this localhost" string
check inline:

- push.go: pushLocal (caddy.Load, recover, forceReload=false) and
  pushRemoteAdmin (POST to the admin /load endpoint) as two small funcs.
- loader.go: a localServer const replaces the scattered "localhost"
  literal; updateServer dispatches on server == localServer.
- generator no longer emits "localhost" into controlledServers (which now
  lists only remote servers); the local target is added in update() when
  this instance also serves.

No behavior change: the local instance still loads in-process, remotes
still POST to the admin API.
